requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.10
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2.11
django-cors-headers==3.11.0
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.1.1
Pillow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
zope.interface==5.4.0
